Fix exit and add join/exit messages
The /exit command should correctly disconnect the node now. Added messages to tell
other clients when a new client joins or leaves the network. Added messages on the
server that echos the clients that join or leave. Fixed a bug with the test file,
so tests should pass. Updated command.ex to have a stronger command pattern.